smart_contract/requirements.txt
beaker-pyteal==1.1.1
pyteal==0.24.1
python-dotenv==1.0.0
